+++ Fabric_Loader/fabric-1.19/src/main/java/com/naptien/managers/RewardEffectManager.java
@@ -1,3 +1,4 @@
+// v5.5.5 Part 73: Fix displayClientMessage in 1.19
 package com.naptien.managers;
 
 import com.naptien.PayBotMod;